+++ libs/apps/uesio/builder/bundle/componentpacks/main/src/components/mainwrapper/canvas.tsx
@@ -75,6 +75,9 @@ const Canvas: FunctionComponent<definition.UtilityProps> = (props) => {
 	}
 
 	const onClickCapture = (e: MouseEvent) => {
+		if (e.shiftKey) {
+			return
+		}
 		e.stopPropagation()
 		e.preventDefault()
